+++ src/main/java/subway/ui/line/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.ui.line.dto;
 
 import subway.application.line.dto.LineDto;
